+++ src/accessory/abstract/base.ts
@@ -14,16 +14,16 @@ import getVersion from '../../tools/version.js';
 
 export abstract class BaseAccessory<C extends BaseAccessoryConfig = BaseAccessoryConfig> extends MQTTAccessory<C> {
 
-  constructor(Service: ServiceType, Characteristic: CharacteristicType, accessory: PlatformAccessory, config: C, log: Log, className: string) {
+  constructor(Service: ServiceType, Characteristic: CharacteristicType, accessory: PlatformAccessory, config: C, log: Log) {
     super(Service, Characteristic, accessory, config, log);
 
     accessory.getService(Service.AccessoryInformation)!
       .setCharacteristic(Characteristic.Name, config.info.name)
       .setCharacteristic(Characteristic.ConfiguredName, config.info.name)
-      .setCharacteristic(Characteristic.Manufacturer, config.info.manufacturer ?? 'Homebridge')
-      .setCharacteristic(Characteristic.SerialNumber, config.info.serialNumber ?? `${PLATFORM_NAME}:${config.info.name}`)
-      .setCharacteristic(Characteristic.Model, config.info.model ?? className)
-      .setCharacteristic(Characteristic.FirmwareRevision, config.info.version ?? getVersion());
+      .setCharacteristic(Characteristic.Manufacturer, PLATFORM_NAME)
+      .setCharacteristic(Characteristic.Model, config.info.type)
+      .setCharacteristic(Characteristic.SerialNumber, config.info.id)
+      .setCharacteristic(Characteristic.FirmwareRevision, getVersion());
 
     this.set(CharacteristicKey.StatusActive, true);